fix: allocator race
