+++ pkg/mem/span.go
@@ -1,0 +1,104 @@
+/*
+ * Copyright 2024 CloudWeGo Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package mem
+
+import (
+	"math/bits"
+	"sync/atomic"
+)
+
+const (
+	spanCacheSize = 10
+	minSpanObject = 128                                  // 128 B
+	maxSpanObject = (minSpanObject << spanCacheSize) - 1 // 128 KB
+	minSpanClass  = 8                                    // = spanClass(minSpanObject)
+)
+
+type spanCache struct {
+	spans [spanCacheSize]*span
+}
+
+func NewSpanCache(spanSize int) *spanCache {
+	c := new(spanCache)
+	for i := 0; i < len(c.spans); i++ {
+		c.spans[i] = NewSpan(spanSize)
+	}
+	return c
+}
+
+func (c *spanCache) Make(n int) []byte {
+	sclass := spanClass(n) - minSpanClass
+	if sclass < 0 || sclass >= len(c.spans) {
+		return make([]byte, n)
+	}
+	return c.spans[sclass].Make(n)
+}
+
+func (c *spanCache) Copy(buf []byte) (p []byte) {
+	p = c.Make(len(buf))
+	copy(p, buf)
+	return p
+}
+
+func NewSpan(size int) *span {
+	sp := new(span)
+	sp.size = uint32(size)
+	sp.buffer = make([]byte, 0, size)
+	return sp
+}
+
+type span struct {
+	lock   uint32
+	read   uint32 // read index of buffer
+	size   uint32 // size of buffer
+	buffer []byte
+}
+
+func (b *span) Make(_n int) []byte {
+	n := uint32(_n)
+	if n >= b.size || !atomic.CompareAndSwapUint32(&b.lock, 0, 1) {
+		// fallback path: make a new byte slice if current goroutine cannot get the lock or n is out of size
+		return make([]byte, n)
+	}
+START:
+	b.read += n
+	// fast path
+	if b.read <= b.size {
+		buf := b.buffer[b.read-n : b.read]
+		atomic.StoreUint32(&b.lock, 0)
+		return buf
+	}
+	// slow path: create a new buffer
+	b.buffer = make([]byte, b.size)
+	b.read = 0
+	goto START
+}
+
+func (b *span) Copy(buf []byte) (p []byte) {
+	p = b.Make(len(buf))
+	copy(p, buf)
+	return p
+}
+
+// spanClass calc the minimum number of bits required to represent x
+// [2^sclass,2^(sclass+1)) bytes in a same span class
+func spanClass(size int) int {
+	if size == 0 {
+		return 0
+	}
+	return bits.Len(uint(size))
+}
