+++ pkg/protocol/bthrift/binary.go
@@ -25,14 +25,17 @@ import (
 
 	"github.com/apache/thrift/lib/go/thrift"
 
+	"github.com/cloudwego/kitex/pkg/mem"
 	"github.com/cloudwego/kitex/pkg/remote/codec/perrors"
 	"github.com/cloudwego/kitex/pkg/utils"
 )
 
-// Binary protocol for bthrift.
-var Binary binaryProtocol
-
-var _ BTProtocol = binaryProtocol{}
+var (
+	// Binary protocol for bthrift.
+	Binary    binaryProtocol
+	_         BTProtocol = binaryProtocol{}
+	spanCache            = mem.NewSpanCache(1024 * 1024) // 1MB
+)
 
 type binaryProtocol struct{}
 
@@ -458,8 +461,7 @@ func (binaryProtocol) ReadString(buf []byte) (value string, length int, err erro
 	if size < 0 || int(size) > len(buf) {
 		return value, length, perrors.NewProtocolErrorWithType(thrift.INVALID_DATA, "[ReadString] the string size greater than buf length")
 	}
-	data := BSpan.Make(int(size))
-	copy(data, buf[length:length+int(size)])
+	data := spanCache.Copy(buf[length : length+int(size)])
 	value = utils.SliceByteToString(data)
 	length += int(size)
 	return
@@ -475,8 +477,7 @@ func (binaryProtocol) ReadBinary(buf []byte) (value []byte, length int, err erro
 	if size < 0 || int(size) > len(buf) {
 		return value, length, perrors.NewProtocolErrorWithType(thrift.INVALID_DATA, "[ReadBinary] the binary size greater than buf length")
 	}
-	value = BSpan.Make(int(size))
-	copy(value, buf[length:length+int(size)])
+	value = spanCache.Copy(buf[length : length+int(size)])
 	length += int(size)
 	return
 }
